btrfs-progs: move extent_io.c to kernel-shared/
