+++ kernel-shared/extent_io.c
@@ -24,7 +24,7 @@
 #include <unistd.h>
 #include <stdbool.h>
 #include "kerncompat.h"
-#include "extent_io.h"
+#include "kernel-shared/extent_io.h"
 #include "kernel-lib/list.h"
 #include "ctree.h"
 #include "volumes.h"
