+++ kernel-shared/free-space-cache.c
@@ -21,7 +21,7 @@
 #include "kernel-shared/free-space-cache.h"
 #include "transaction.h"
 #include "disk-io.h"
-#include "extent_io.h"
+#include "kernel-shared/extent_io.h"
 #include "crypto/crc32c.h"
 #include "kernel-lib/bitops.h"
 #include "common/internal.h"
